+++ src/models/deviceSettings/deviceSettings.ts
@@ -5,7 +5,7 @@ import { Utils } from '../../server';
 export abstract class DeviceSettings extends ObjectSettings {
   public energySettings: ExcessEnergyConsumerSettings | undefined = undefined;
 
-  public fromPartialObject(_obj: Partial<DeviceSettings>): void {
+  public override fromPartialObject(_obj: Partial<DeviceSettings>): void {
     if (_obj.energySettings) {
       if (this.energySettings === undefined) {
         this.energySettings = new ExcessEnergyConsumerSettings();