+++ src/models/deviceSettings/ledSettings.ts
@@ -2,7 +2,8 @@ import { DimmerSettings } from './dimmerSettings';
 import { Utils } from '../../server';
 
 export class LedSettings extends DimmerSettings {
-  public defaultColor: string = '#fbbc32';
+  public static fallbackColor: string = '#fbbc32';
+  public defaultColor: string = LedSettings.fallbackColor;
   public dayOn: boolean = false;
   public dayBrightness: number = 100;
   public dayColor: string = this.defaultColor;
@@ -20,24 +21,24 @@ export class LedSettings extends DimmerSettings {
   public nightColor: string = '#ff5500';
   public nightColorTemp: number = -1;
 
-  public fromPartialObject(data: Partial<LedSettings>): void {
-    this.defaultColor = data.defaultColor ?? this.defaultColor;
+  public override fromPartialObject(data: Partial<LedSettings>): void {
     this.dayOn = data.dayOn ?? this.dayOn;
     this.dayBrightness = data.dayBrightness ?? this.dayBrightness;
-    this.dayColor = data.dayColor ?? this.dayColor;
     this.dayColorTemp = data.dayColorTemp ?? this.dayColorTemp;
     this.dawnOn = data.dawnOn ?? this.dawnOn;
     this.dawnBrightness = data.dawnBrightness ?? this.dawnBrightness;
-    this.dawnColor = data.dawnColor ?? this.dawnColor;
     this.dawnColorTemp = data.dawnColorTemp ?? this.dawnColorTemp;
     this.duskOn = data.duskOn ?? this.duskOn;
     this.duskBrightness = data.duskBrightness ?? this.duskBrightness;
-    this.duskColor = data.duskColor ?? this.duskColor;
     this.duskColorTemp = data.duskColorTemp ?? this.duskColorTemp;
     this.nightOn = data.nightOn ?? this.nightOn;
     this.nightBrightness = data.nightBrightness ?? this.nightBrightness;
-    this.nightColor = data.nightColor ?? this.nightColor;
     this.nightColorTemp = data.nightColorTemp ?? this.nightColorTemp;
+    this.defaultColor = Utils.formatHex(data.defaultColor ?? this.defaultColor) ?? LedSettings.fallbackColor;
+    this.dayColor = Utils.formatHex(data.dayColor ?? this.dayColor) ?? LedSettings.fallbackColor;
+    this.dawnColor = Utils.formatHex(data.dawnColor ?? this.dawnColor) ?? LedSettings.fallbackColor;
+    this.duskColor = Utils.formatHex(data.duskColor ?? this.duskColor) ?? LedSettings.fallbackColor;
+    this.nightColor = Utils.formatHex(data.nightColor ?? this.nightColor) ?? LedSettings.fallbackColor;
     super.fromPartialObject(data);
   }